maj persistence

+++ vol-persistence/src/main/java/vol/metier/dao/impl/AeroportDaoJpa.java
@@ -12,10 +12,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import vol.metier.dao.AeroportDao;
 import vol.metier.dao.EscaleDao;
+import vol.metier.dao.VolDao;
 import vol.metier.dao.AeroportVilleDao;
 import vol.metier.model.Aeroport;
 import vol.metier.model.AeroportVille;
 import vol.metier.model.Escale;
+import vol.metier.model.Vol;
 
 @Transactional
 @Repository
@@ -29,6 +31,9 @@ public class AeroportDaoJpa implements AeroportDao {
 
 	@Autowired
 	private EscaleDao escaleDao;
+	@Autowired
+	private VolDao volDao;
+	
 
 	@Override
 	@Transactional(readOnly=true)
@@ -66,16 +71,31 @@ public class AeroportDaoJpa implements AeroportDao {
 
 	@Override
 	public void delete(Aeroport aeroport) {
-		em.refresh(aeroport);
-		for (AeroportVille villeAeroport : aeroport.getVilles()) {
-			aeroportVilleDao.delete(villeAeroport);
+		System.out.println("etapa 1 delete en AeroportJpa");
+		em.merge(aeroport);
+		
+		//
+		List<Vol> LV = volDao.findAll();
+		for (int i = 0; i < LV.size(); i++) {
+			
+			if (LV.get(i).getDepart().equals(aeroport) || LV.get(i).getArrivee().equals(aeroport))
+			{System.out.println("ETAPA EN XXXXXX");
+				volDao.delete(volDao.find(LV.get(i).getId()));
+			}
 		}
-		for (Escale escale : aeroport.getEscales()) {
+		 //
+		if(aeroport.getVilles().size()!=0){System.out.println("aeroport.getVilles().size()!=0 en AeroDaoJpa");
+			for (AeroportVille villeAeroport : aeroport.getVilles()) {System.out.println("etapa 2 delete en AeroportJpa");
+				aeroportVilleDao.delete(villeAeroport);
+			}
+		}
+		
+		for (Escale escale : aeroport.getEscales()) {System.out.println("etapa 3 delete en AeroportJpa");
 			escaleDao.delete(escale);
 		}
-
-		em.remove(aeroport);
-
+		System.out.println("etapa 4 antes remove en AeroportJpa");
+		em.remove(em.contains(aeroport) ? aeroport : em.merge(aeroport));
+		System.out.println("saliendo bien de AeroportJpa");
 	}
